requirements.txt
uvicorn[standard]
google-cloud-bigquery
google-cloud-firestore
google-auth
singleton-decorator
firebase-admin
google-auth
pandas
pyarrow
iteration-utilities
